cp: replace pre-calculated cumulative with in-code one.
The cumulative used to be entered manually, pre-calculated
out-of-code. This was an artifact of the initial script by mic.

(Why it held on for so long is a question to ffwd...)

This change allows for the direct entry of a daily case number;
the cumulative is then calculated on the fly.

To transform one list to another, this was used:

DAILY_CASES = []
previous = 0
for cumulative in CUMULATIVE_CASES:
    DAILY_CASES.append(cumulative - previous)
    previous = cumulative

To transform back (as currently in code):

cumulative = 0
CUMULATIVE_CASES = []
for day_cases in DAILY_CASES:
    cumulative += day_cases
    CUMULATIVE_CASES.append(cumulative)

If you know how to do this in a single line, send me an e-mail
or something after the hangover is gone.